+++ acme/resources/CNT_LA.py
@@ -25,12 +25,12 @@ class CNT_LA(Resource):
 		return super()._canHaveChild(resource, [])
 
 
-	def asJSON(self):	# overwrite
+	def asJSON(self, embedded=True, update=False, noACP=False):
 		pi = self['pi']
 		Logging.logDebug('Latest CIN from CNT: %s' % pi)
 		(pr, _) = CSE.dispatcher.retrieveResource(pi)	# get parent
 		rs = pr.contentInstances()						# ask parent for all CIN
 		if len(rs) == 0:								# In case of none
 			return None
-		return rs[-1].asJSON()							# result is sorted, so take, and return last
+		return rs[-1].asJSON(embedded=embedded, update=update, noACP=noACP)		# result is sorted, so take, and return last