+++ acme/resources/CNT_OL.py
@@ -24,11 +24,11 @@ class CNT_OL(Resource):
 		return super()._canHaveChild(resource, [])
 
 
-	def asJSON(self):	# overwrite
+	def asJSON(self, embedded=True, update=False, noACP=False):
 		pi = self['pi']
 		Logging.logDebug('Oldest CIN from CNT: %s' % pi)
 		(pr, _) = CSE.dispatcher.retrieveResource(pi)	# get parent
 		rs = pr.contentInstances()						# ask parent for all CIN
 		if len(rs) == 0:								# In case of none
 			return None
-		return rs[0].asJSON()							# result is sorted, so take, and return first
+		return rs[0].asJSON(embedded=embedded, update=update, noACP=noACP)		# result is sorted, so take, and return first