+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 - [CSE] Discovery supports "attributes + children" return content.
 - [CSE] Changed command line argument --reset-db to --db-reset .
 - [CSE] Added command line argument --db-storage .
-- [CSE] Added support for FlexContainerInstance. 
+- [CSE] Added support for FlexContainerInstance.
+- [CSE] Fixed discovery results: ignore latest, oldest, and fixed result format.
 
 ## 0.2.1 - 2020-03-06
 - [APPS] Fixed wrong originator handling for already registered AEs.